Squashed commit of the following:
commit c28442a7ab3834cf0cbf5347c7935fc45f1dab65
Author: Gagan Deep <pandafy.dev@gmail.com>
Date:   Thu Aug 7 17:36:01 2025 +0530

    [tests] Fixed failing tests

commit c92793d97cb227dc3398f97f572fa015a4a7abee
Author: Gagan Deep <pandafy.dev@gmail.com>
Date:   Tue Jul 29 22:55:22 2025 +0530

    [fix:ux] Fixed rendering of JSON fields in readonly mode #848

    Fixes #848

commit 3817db7815a68c240defe5679048bf4341e37237
Author: Gagan Deep <pandafy.dev@gmail.com>
Date:   Sat Aug 2 03:35:22 2025 +0530

    [docs] Added changelog for Version 1.1.2

commit 6d997e10803a752d512019263a674073c597af31
Author: Gagan Deep <pandafy.dev@gmail.com>
Date:   Wed Jul 30 21:05:13 2025 +0530

    [fix] Don't delete VpnClients when enforcing required templates

    Bug:
    The SortedManyToManyField clears all templates before adding new ones.
    This operation emits an `m2m_changed` signal with the `post_clear` action,
    which triggers the `Config.enforce_required_templates` receiver to add
    required templates back.

    Adding required templates emits another `m2m_changed` signal with the
    `post_add` action, which executes `Config.manage_vpn_clients` receiver.
    At this point, only required templates exist in the DB, hence we cannot
    decide which VpnClient objects to delete.

    Fix:
    We don't delete any VpnClient objects when the `m2m_changed` signal
    is sent for the `post_add` action for adding required templates.
    The receiver will be called again with the `post_add` action when
    all the templates are added back, including the required ones.
    And then, it will delete any VpnClient objects that are not
    associated with the current templates.

    ---------


commit f9f0250ffb851466a3a30b4860e99b0ccd23f398
Author: Federico Capoano <f.capoano@openwisp.io>
Date:   Wed Jul 30 11:22:38 2025 -0300

    [chores:tests] Moved config.tests.test_admin.tearDownClass up

commit 6ae8103d4b4eaf1c0d1f9e4dc2f6f93a3c6953f0
Author: Federico Capoano <f.capoano@openwisp.io>
Date:   Tue Jul 29 13:49:39 2025 -0300

    [fix] Use uuid path converter in geo URLs

commit 3b7ff66edaf830075a28a4c12ddb6b17908d55df
Author: Gagan Deep <pandafy.dev@gmail.com>
Date:   Tue Jul 29 20:09:25 2025 +0530

    [docs/ui] Improve docs of send commands API endpoint #1047

    Closes #1047

commit 6b33d7304cf017c9623bfa6b101fe009b6f00acd
Author: Gagan Deep <pandafy.dev@gmail.com>
Date:   Mon Jul 28 20:13:24 2025 +0530

    [feature] Cache VPN server checksum #1064

    Closes #1064

commit d91e11d414f2f8a762811a2fffb7d9fb98320d40
Author: Gagan Deep <pandafy.dev@gmail.com>
Date:   Tue Jul 22 21:36:42 2025 +0530

    [fix] Fixed visibility of "Recent Commands" tab on device page

    [skip changelog]

commit c9463ba53a7355890570742c5540b10b80a653fc
Author: AlexV <asmodehn@gmail.com>
Date:   Tue Jul 22 17:45:26 2025 +0200

    [tests] Enabled redis caching and session storage in automated tests

    * [change] Sessions now cached in redis
    * [fix] Adjust count in config.tests.TestAdmin._verify_template_queries

    ---------


commit 4ea68be945a3b7fc15146b9d47a79f315d5f682d
Author: Federico Capoano <f.capoano@openwisp.io>
Date:   Thu Jul 3 19:04:46 2025 -0300

    [fix] Fixed 0060_cleanup_api_task_notification_types (missing dep)

    This migration lacked an explicit dependency on the notifications
    module's migrations, which could cause it to run before the
    NotificationSetting model is available, resulting in a crash.

+++ openwisp_controller/config/base/config.py
@@ -3,7 +3,6 @@ import logging
 import re
 from collections import defaultdict
 
-from cache_memoize import cache_memoize
 from django.core.exceptions import ObjectDoesNotExist, PermissionDenied, ValidationError
 from django.db import models, transaction
 from django.utils.translation import gettext_lazy as _
@@ -24,9 +23,18 @@ from ..signals import (
 )
 from ..sortedm2m.fields import SortedManyToManyField
 from ..utils import get_default_templates_queryset
-from .base import BaseConfig
+from .base import BaseConfig, ChecksumCacheMixin, get_cached_args_rewrite
 
 logger = logging.getLogger(__name__)
+
+
+def get_cached_checksum_args_rewrite(config):
+    """
+    DEPRECATED: Use get_cached_args_rewrite instead.
+
+    TODO: Remove this in 1.2.0 release.
+    """
+    return get_cached_args_rewrite(config)
 
 
 class TemplatesThrough(object):
@@ -38,14 +46,7 @@ class TemplatesThrough(object):
         return _("Relationship with {0}").format(self.template.name)
 
 
-def get_cached_checksum_args_rewrite(config):
-    """
-    Use only the PK parameter for calculating the cache key
-    """
-    return config.pk.hex
-
-
-class AbstractConfig(BaseConfig):
+class AbstractConfig(ChecksumCacheMixin, BaseConfig):
     """
     Abstract model implementing the
     NetJSON DeviceConfiguration object
@@ -149,23 +150,6 @@ class AbstractConfig(BaseConfig):
         (kept for backward compatibility with pre 0.6 versions)
         """
         return self.device.key
-
-    @cache_memoize(
-        timeout=_CHECKSUM_CACHE_TIMEOUT, args_rewrite=get_cached_checksum_args_rewrite
-    )
-    def get_cached_checksum(self):
-        """
-        Handles caching,
-        timeout=None means value is cached indefinitely
-        (invalidation handled on post_save/post_delete signal)
-        """
-        logger.debug(f"calculating checksum for config ID {self.pk}")
-        return self.checksum
-
-    @classmethod
-    def bulk_invalidate_get_cached_checksum(cls, query_params):
-        for config in cls.objects.only("id").filter(**query_params).iterator():
-            config.get_cached_checksum.invalidate(config)
 
     @classmethod
     def get_template_model(cls):
@@ -330,10 +314,27 @@ class AbstractConfig(BaseConfig):
         else:
             templates = pk_set
 
-        # Delete VPN clients that are not associated with current templates
-        instance.vpnclient_set.exclude(
-            template_id__in=instance.templates.values_list("id", flat=True)
-        ).delete()
+        # Check if all templates in pk_set are required templates. If they are,
+        # skip deletion of VpnClient objects at this point.
+        if len(pk_set) != templates.filter(required=True).count():
+            # Explanation:
+            # SortedManyToManyField clears all existing templates before adding
+            # new ones. This triggers an m2m_changed signal with the "post_clear"
+            # action, which is handled by the "enforce_required_templates" signal
+            # receiver. That receiver re-adds the required templates.
+            #
+            # Re-adding required templates triggers another m2m_changed signal
+            # with the "post_add" action. At this stage, only required templates
+            # exist in the DB, so we cannot yet determine which VpnClient objects
+            # should be deleted based on the new selection.
+            #
+            # Therefore, we defer deletion of VpnClient objects until the "post_add"
+            # signal is triggered again—after all templates, including the required
+            # ones, have been fully added. At that point, we can identify and
+            # delete VpnClient objects not linked to the final template set.
+            instance.vpnclient_set.exclude(
+                template_id__in=instance.templates.values_list("id", flat=True)
+            ).delete()
 
         if action == "post_add":
             for template in templates.filter(type="vpn"):
